Move grid_state functionallity to a base class for XLS report views.
